examples/requirements.txt
langchain-core>=0.3.35
langgraph~=0.3.5
pydantic==2.10.6
langchain-openai==0.3.6
langchain>=0.3.19
langchain-community==0.3.19
faiss-cpu==1.10.0
llama-index-core==0.12.22
llama-index-utils-workflow==0.3.0
llama-index-llms-azure-openai==0.3.1
llama-index-vector-stores-docarray>=0.3.0
llama-index-embeddings-huggingface>=0.5.2
-e ../
